Redo the dependency handling from scratch to make it more flexible
With the new system, one can either declare installation information for a
dependency inline in the call to socute_add_package, or inside a package file
module that can declare variables and macros to define how a package should
be found and installed.

The package file modules can now be placed anywhere, the additional places
must be declared with a call to socute_add_package_module_dir(). Those directories
can either contain a module with the name of the package, or a module template
whose name should be a prefix to the package to be found/installed. These
templates can be used to factorize code for a number of similar packages in a
unique file.

Documentation to come later.