Add a new HEADERS option to socute_find_package for single file libraries
Declaring one or several files in the HEADERS options field allows the use
of self contained headers-only libraries as import libraries.
The headers will be fetched and installed in an include directory, and an
import lib will be created.

The HEADERS_DIR option can be used to constrain the installation of the
downloaded headers in a specific subdirectory.